+++ packages/eyes-images/test/it/TestFluentApi.spec.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const { ConsoleLogHandler, Region, RectangleSize } = require('@applitools/eyes-sdk-core');
-const { Eyes, Target } = require('../../index');
+const { Eyes, Target, ConsoleLogHandler, Region, RectangleSize } = require('../../index');
 
 let /** @type {Eyes} */ eyes;
 describe('EyesImages.TestFluentApi', function () {
@@ -47,6 +46,6 @@ describe('EyesImages.TestFluentApi', function () {
   });
 
   afterEach(async function () {
-    await eyes.abortIfNotClosed();
+    await eyes.abort();
   });
 });