+++ packages/eyes-sdk-core/lib/TestResultsFormatter.js
@@ -18,7 +18,7 @@ class TestResultsFormatter {
   /**
    * Adds an additional results object to the currently stored results list.
    *
-   * @param {TestResults} results - A test results returned by a call to `eyes.close' or 'eyes.abortIfNotClosed'.
+   * @param {TestResults} results - A test results returned by a call to `eyes.close' or 'eyes.abort'.
    * @return {TestResultsFormatter} - The updated 'TestResultsFormatter' instance.
    */
   addTestResults(results) {
@@ -34,7 +34,7 @@ class TestResultsFormatter {
    * Adds an additional results object to the currently stored results list.
    *
    * @deprecated use {@link #addTestResults(results)} instead
-   * @param {TestResults} results - A test results returned by a call to `eyes.close' or 'eyes.abortIfNotClosed'.
+   * @param {TestResults} results - A test results returned by a call to `eyes.close' or 'eyes.abort'.
    * @return {TestResultsFormatter} - The updated 'TestResultsFormatter' instance.
    */
   addResults(results) {
@@ -152,9 +152,6 @@ class TestResultsFormatter {
           }
           tapString += `#\tMismatches: ${currentResult.getMismatches()}\n`;
         }
-        if (currentResult.getIsSaved()) {
-          tapString += '#\tTest was automatically saved as a baseline.\n';
-        }
       }
 
       const url = currentResult.getAppUrls() && currentResult.getAppUrls().getSession() ?
@@ -221,9 +218,6 @@ class TestResultsFormatter {
             tapString += `[FAILED TEST] ${name}\n`;
           }
           tapString += `#\tMismatches: ${currentResult.getMismatches()}\n`;
-        }
-        if (currentResult.getIsSaved()) {
-          tapString += '#\tTest was automatically saved as a baseline.\n';
         }
       }