+++ packages/eyes-selenium/lib/fluent/SeleniumCheckSettings.js
@@ -40,7 +40,7 @@ class SeleniumCheckSettings extends CheckSettings {
       this.frame(frame);
     }
 
-    /** @type {Map<string, string>} */ this._scriptHooks = new Map();
+    /** @type {Object.<string, string>} */ this._scriptHooks = {};
   }
 
   /**
@@ -294,17 +294,26 @@ class SeleniumCheckSettings extends CheckSettings {
   }
 
   /**
+   * @deprecated
    * @param {String} hook
    * @return {this}
    */
   webHook(hook) {
-    this._scriptHooks.set(BEFORE_CAPTURE_SCREENSHOT, hook);
+    return this.beforeRenderScreenshotHook(hook);
+  }
+
+  /**
+   * @param {String} hook
+   * @return {this}
+   */
+  beforeRenderScreenshotHook(hook) {
+    this._scriptHooks[BEFORE_CAPTURE_SCREENSHOT] = hook;
     return this;
   }
 
   /**
    * @ignore
-   * @return {Map<string, string>}
+   * @return {Object.<string, string>}
    */
   getScriptHooks() {
     return this._scriptHooks;