+++ packages/eyes-images-browser/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.28](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.27...@applitools/eyes-images-browser@4.6.28) (2019-07-03)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.27](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.26...@applitools/eyes-images-browser@4.6.27) (2019-07-02)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.26](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.25...@applitools/eyes-images-browser@4.6.26) (2019-06-30)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.25](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.24...@applitools/eyes-images-browser@4.6.25) (2019-06-24)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.24](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.23...@applitools/eyes-images-browser@4.6.24) (2019-06-19)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.23](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.22...@applitools/eyes-images-browser@4.6.23) (2019-06-19)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
+## [4.6.22](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.21...@applitools/eyes-images-browser@4.6.22) (2019-06-17)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
 ## [4.6.21](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.20...@applitools/eyes-images-browser@4.6.21) (2019-06-03)
 
 **Note:** Version bump only for package @applitools/eyes-images-browser